sm8350-common: parts: sepolicy: Initial Sepolicy for Xiaomi parts
Change-Id: 68fed5cda7f4c17aab13ga47e6e212b97b22c2bf
Signed-off-by: kenway214 <kenway214@outlook.com>

peridot: sepolicy: allow parts to get SettingsLib prop

ThermalSettingsFragment's use of ApplicationsState includes a
check of whether a package is an instant app or not. The function
isInstant() in AppUtils reads settingsdebug.instant.packages prop
to retrieve a list of instant packages for debugging purposes.

We do not actually use this property, but this check triggers an
SELinux denial for every package in the list and spams the logs
with permission denied errors. Get rid of these log spams by
allowing parts to read the prop.

Test: open the thermal profile settings and check logs
Signed-off-by: kenway214 <kenway214@outlook.com>
Co-Authored-By: Arian <arian.kulmer@web.de>